requirements.txt
# Core IB API Dependencies
ib_async==1.0.3

# Data Processing
pandas>=2.0.0
pandas-market-calendars>=4.0.0
pytz>=2023.3

# Configuration & Storage
PyYAML>=6.0.1
python-dateutil>=2.8.2

# Logging & Monitoring
rich>=13.0.0

# Development Tools
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Additional utilities
asyncio-throttle>=1.0.2  # For rate limiting
nest-asyncio>=1.5.6      # For nested event loops if needed 